% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\docType{class}
\name{AnyHtmlWidget}
\alias{AnyHtmlWidget}
\title{AnyHtmlWidget Class}
\value{
The value.

The ESM string.

List of values.

The width.

The height.

The port number.
}
\description{
Class representing a widget.
}
\details{
AnyHtmlWidget
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnyHtmlWidget-new}{\code{AnyHtmlWidget$new()}}
\item \href{#method-AnyHtmlWidget-set_value}{\code{AnyHtmlWidget$set_value()}}
\item \href{#method-AnyHtmlWidget-on_change}{\code{AnyHtmlWidget$on_change()}}
\item \href{#method-AnyHtmlWidget-get_value}{\code{AnyHtmlWidget$get_value()}}
\item \href{#method-AnyHtmlWidget-get_esm}{\code{AnyHtmlWidget$get_esm()}}
\item \href{#method-AnyHtmlWidget-get_values}{\code{AnyHtmlWidget$get_values()}}
\item \href{#method-AnyHtmlWidget-get_width}{\code{AnyHtmlWidget$get_width()}}
\item \href{#method-AnyHtmlWidget-get_height}{\code{AnyHtmlWidget$get_height()}}
\item \href{#method-AnyHtmlWidget-set_values}{\code{AnyHtmlWidget$set_values()}}
\item \href{#method-AnyHtmlWidget-set_mode}{\code{AnyHtmlWidget$set_mode()}}
\item \href{#method-AnyHtmlWidget-start_server}{\code{AnyHtmlWidget$start_server()}}
\item \href{#method-AnyHtmlWidget-stop_server}{\code{AnyHtmlWidget$stop_server()}}
\item \href{#method-AnyHtmlWidget-get_host}{\code{AnyHtmlWidget$get_host()}}
\item \href{#method-AnyHtmlWidget-get_port}{\code{AnyHtmlWidget$get_port()}}
\item \href{#method-AnyHtmlWidget-print}{\code{AnyHtmlWidget$print()}}
\item \href{#method-AnyHtmlWidget-render}{\code{AnyHtmlWidget$render()}}
\item \href{#method-AnyHtmlWidget-clone}{\code{AnyHtmlWidget$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-new}{}}}
\subsection{Method \code{new()}}{
Create a new widget instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$new(.esm, .mode, .width = NA, .height = NA, .commands = NA, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.esm}}{The EcmaScript module as a string.}

\item{\code{.mode}}{The widget mode.}

\item{\code{.width}}{The widget width. Optional.}

\item{\code{.height}}{The widget height. Optional.}

\item{\code{.commands}}{TODO}

\item{\code{...}}{All other named arguments will be used to create active bindings on the instance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-set_value"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-set_value}{}}}
\subsection{Method \code{set_value()}}{
Set a value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$set_value(key, val, emit_change = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The key of the value to set.}

\item{\code{val}}{The new value.}

\item{\code{emit_change}}{Should the on_change handler be called?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-on_change"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-on_change}{}}}
\subsection{Method \code{on_change()}}{
Register a change handler to call if emit_change is TRUE in set_value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$on_change(callback)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{callback}}{A callback function to register.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-get_value}{}}}
\subsection{Method \code{get_value()}}{
Get a particular value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$get_value(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The key of the value to get.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-get_esm"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-get_esm}{}}}
\subsection{Method \code{get_esm()}}{
Get the ESM string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$get_esm()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-get_values"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-get_values}{}}}
\subsection{Method \code{get_values()}}{
Get all widget values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$get_values()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-get_width"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-get_width}{}}}
\subsection{Method \code{get_width()}}{
Get the widget width.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$get_width()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-get_height"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-get_height}{}}}
\subsection{Method \code{get_height()}}{
Get the widget height.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$get_height()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-set_values"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-set_values}{}}}
\subsection{Method \code{set_values()}}{
Set all values. TODO: is this ever used?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$set_values(new_values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_values}}{A list of new values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-set_mode"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-set_mode}{}}}
\subsection{Method \code{set_mode()}}{
Set the widget mode.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$set_mode(mode)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mode}}{The new widget mode.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-start_server"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-start_server}{}}}
\subsection{Method \code{start_server()}}{
Start the server, if not running.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$start_server()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-stop_server"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-stop_server}{}}}
\subsection{Method \code{stop_server()}}{
Stop the server, if running.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$stop_server()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-get_host"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-get_host}{}}}
\subsection{Method \code{get_host()}}{
Get the server hostname.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$get_host()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The hostname as a string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-get_port"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-get_port}{}}}
\subsection{Method \code{get_port()}}{
Get the server port.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$get_port()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-print"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-print}{}}}
\subsection{Method \code{print()}}{
Custom print function for the R6 class.
If mode is "shiny", falls back to original R6 print behavior.
Otherwise, renders the widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-render"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-render}{}}}
\subsection{Method \code{render()}}{
Render the widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$render()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyHtmlWidget-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnyHtmlWidget-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyHtmlWidget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
