% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{the_anyhtmlwidget}
\alias{the_anyhtmlwidget}
\title{The internal function that creates the htmlwidget.}
\usage{
the_anyhtmlwidget(
  esm,
  values = NULL,
  ns_id = NULL,
  width = NULL,
  height = NULL,
  port = NULL,
  host = NULL,
  element_id = NULL
)
}
\arguments{
\item{esm}{The ES Module as a string.}

\item{values}{The values that will be used for the initial Model state.}

\item{ns_id}{Namespace ID, only used when in the Shiny module mode. Optional.}

\item{width}{The width of the widget as a number or CSS string. Optional.}

\item{height}{The height of the widget as a number or CSS string. Optional.}

\item{port}{The port of the WebSocket server, when in dynamic mode. Optional.}

\item{host}{The host of the WebSocket server, when in dynamic mode. Optional.}

\item{element_id}{An element ID. Optional.}
}
\value{
The result of htmlwidgets::createWidget.
}
\description{
The internal function that creates the htmlwidget.
}
\keyword{internal}
