% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{widgetServer}
\alias{widgetServer}
\title{Shiny module server for anyhtmlwidgets.}
\usage{
widgetServer(id, w)
}
\arguments{
\item{id}{The matching output ID used in the Shiny UI.}

\item{w}{The widget instance.}
}
\value{
reactiveValues corresponding to the widget's active bindings.
}
\description{
Shiny module server for anyhtmlwidgets.
}
