% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{widgetUI}
\alias{widgetUI}
\title{Shiny module UI for anyhtmlwidgets.}
\usage{
widgetUI(id, width = "100\%", height = "400px")
}
\arguments{
\item{id}{The output ID.}

\item{width}{The widget width. Optional.}

\item{height}{The widget height. Optional.}
}
\description{
Shiny module UI for anyhtmlwidgets.
}
